Move com.typesafe.sbt.pgp to com.jsuereth.sbtpgp
It's odd for the namespace of the plugin to not match the artifact
namespace.
Also, cleans up obvious, low-hanging-fruit deprecation warnings.
