+++ sbt-pgp/src/main/scala-sbt-0.13/Compat.scala
