+++ sbt-pgp/src/main/scala-sbt-1.0/Compat.scala
