+++ sbt-pgp/src/main/scala/com/jsuereth/pgp/cli/ListSigs.scala
