+++ sbt-pgp/src/main/scala/com/jsuereth/pgp/cli/commands.scala
