+++ sbt-pgp/src/main/scala/com/jsuereth/pgp/cli/context.scala
