+++ sbt-pgp/src/main/scala/com/jsuereth/pgp/cli/hkpcommands.scala
