+++ sbt-pgp/src/main/scala/com/jsuereth/sbtpgp/SbtPgp.scala
@@ -1,11 +1,10 @@
-package com.typesafe.sbt
+package com.jsuereth.sbtpgp
 
-import com.typesafe.sbt.pgp._
 import sbt._
 import sbt.sbtpgp.Compat._
 
 /**
- * This class is used to control what we expose to 
+ * This class is used to control what we expose to
  * users.   It grants access to all our keys in the
  * common naming sense of plugins.   This is temporary
  * until we clean this plugin up for 0.12.0 usage.
@@ -15,10 +14,8 @@ object SbtPgp extends AutoPlugin {
   override def trigger = allRequirements
   override def requires = pgpRequires
 
-  // Note - workaround for issues in sbt 0.13.5 autoImport
-  object autoImportImpl {
-
-    val PgpKeys = pgp.PgpKeys
+  object autoImport {
+    val PgpKeys = com.jsuereth.sbtpgp.PgpKeys
 
     // TODO - Are these ok for style guide?  We think so.
     def useGpg = PgpKeys.useGpg in Global
@@ -36,7 +33,6 @@ object SbtPgp extends AutoPlugin {
 
     def signingSettings = PgpSettings.signingSettings
   }
-  val autoImport = autoImportImpl
   // TODO - Maybe signing settigns should be a different plugin...
   override val projectSettings = PgpSettings.projectSettings
   override val globalSettings = PgpSettings.globalSettings
