+++ sbt-pgp/src/main/scala/com/jsuereth/sbtpgp/SbtPgpCommandContext.scala
@@ -1,5 +1,4 @@
-package com.typesafe.sbt
-package pgp
+package com.jsuereth.sbtpgp
 
 import sbt._
 import sbt.Keys.TaskStreams
