+++ sbt-pgp/src/sbt-test/pgp-cmd/gen-key/project/plugins.sbt
