+++ sbt-pgp/src/sbt-test/pgp-cmd/gen-key/test
