+++ sbt-pgp/src/sbt-test/sbt-pgp/credentials/build.sbt
