+++ sbt-pgp/src/sbt-test/sbt-pgp/credentials/project/plugins.sbt
