+++ sbt-pgp/src/sbt-test/sbt-pgp/credentials/test
