+++ sbt-pgp/src/sbt-test/sbt-pgp/skip/project/plugins.sbt
